Separate out single leader only tests
